--- app/api-service/src/main/java/com/cnu/spg/security/config/SecurityWebApplicationAdapter.java
@@ -1,10 +1,0 @@
-package com.cnu.spg.security.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.web.context.AbstractSecurityWebApplicationInitializer;
-
-@Configuration
-@EnableWebSecurity
-public class SecurityWebApplicationAdapter extends AbstractSecurityWebApplicationInitializer {
-}
